refactor(test): root from arbitrary number of digests belongs to MMR
